My revisions to Phoa section
Sorry for the kind of tumultuous revisions. After re-reading the paper,
I realised that the connection to Phoa's principle was made a bit too
obscurely, and that many of the results also rely only on Phoa's
principle so it was a bit of a shame to require SQCF for them.

To that end, I add SQCP which states that I[i] is quasi-coherent. This
is equivalent to Phoa's principle, but it is also equivalent to the
quasi-coherence principle for *all* finitely generated free I-algebras,
which I prove.

I also develop a bit more fully the comparison between the different
kinds of orders, including path order and observation order, etc.

I am not sure that my revisions are perfect (please check them), but I
think that it is improving the structure of the paper.